+++ client/src/components/common/Selector/Dropdown/Item.tsx
@@ -1,4 +1,4 @@
-import { useDropdownContext } from './DropdownContext';
+import { useDropdownContext } from 'src/hooks/useDropdownContext';
 
 interface ItemProps {
   id: number;
